% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBDInfer.R
\name{global_variables}
\alias{global_variables}
\title{Global Variables for IBDInfer}
\description{
This section declares global variables used in the IBDInfer package to prevent R CMD check warnings.
}
\keyword{internal}
