% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.IBD.R
\name{summary.SS}
\alias{summary.SS}
\title{Summary of IBD}
\usage{
\method{summary}{SS}(object, ...)
}
\value{
No return value, called for summary.
}
\description{
Summary function for IBDInfer
}
\keyword{internal}
